requirements.txt
# SENTINEL-X Python Dependencies
# AutoGen + Groq + ML Pipeline (100% Free Tier)
# Requirements for agents/, ML pipeline, CI/CD

# Core
pyautogen>=0.2.0
groq>=0.5.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utils
python-dotenv>=1.0.0

# Testing (free tier)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Load testing (free tier)
locust>=2.0.0